+++ tracing_test.go
@@ -313,6 +313,22 @@ func TestStartTransaction(t *testing.T) {
 	}
 }
 
+func TestIsTansaction(t *testing.T) {
+	ctx := NewTestContext(ClientOptions{
+		EnableTracing: true,
+	})
+
+	transaction := StartTransaction(ctx, "Test Transaction", func(s *Span) {})
+	if !transaction.IsTransaction() {
+		t.Fatalf("span.IsTransaction() = false, want true")
+	}
+
+	span := transaction.StartChild("Test Span", func(s *Span) {})
+	if span.IsTransaction() {
+		t.Fatalf("span.IsTransaction() = true, want false")
+	}
+}
+
 // testContextKey is used to store a value in a context so that we can check
 // that SDK operations on that context preserve the original context values.
 type testContextKey struct{}