+++ tracing.go
@@ -221,6 +221,11 @@ func (s *Span) SetData(name, value string) {
 	s.Data[name] = value
 }
 
+// IsTransaction checks if the given span is a transaction.
+func (s *Span) IsTransaction() bool {
+	return s.isTransaction
+}
+
 // TODO(tracing): maybe add shortcuts to get/set transaction name. Right now the
 // transaction name is in the Scope, as it has existed there historically, prior
 // to tracing.