+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add SentryTraceHeader and SentryBaggageHeader constants ([#538](https://github.com/getsentry/sentry-go/pull/538/))
 - Add `Span.SetDynamicSamplingContext` method ([#539](https://github.com/getsentry/sentry-go/pull/539/))
 - Add DSN getters ([#540](https://github.com/getsentry/sentry-go/pull/540))
+- Add Span.IsTransaction() ([#543](https://github.com/getsentry/sentry-go/pull/543))
 
 ## 0.17.0